:recycle: [EmailClasses] Reorganise into specific folder
