+++ src/ServiceV2/Mailer/Email/DuplicatedUsernameEmail.php
@@ -1,12 +1,15 @@
 <?php
 
-namespace App\ServiceV2\Mailer;
+namespace App\ServiceV2\Mailer\Email;
 
 use App\Entity\User;
 use Symfony\Component\Mime\Email;
 
 readonly class DuplicatedUsernameEmail
 {
+    /**
+     * @param string[] $recipients
+     */
     public static function create(string $sender, array $recipients, User $duplicatedUser): Email
     {
         return (new Email())
