+++ src/ServiceV2/Mailer/Email/LocalizedTemplatedEmailInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\ServiceV2\Mailer;
+namespace App\ServiceV2\Mailer\Email;
 
 use App\Entity\User;
 use Brevo\Client\Model\SendSmtpEmail;
